Add a float cast to prevent an ambiguous call to sqrt with msvc

+++ libraries/terry/src/terry/generator/colorcube.hpp
@@ -51,7 +51,7 @@ struct ColorCubeFunctor
 			return pixel;
 		}
 		
-		size_t cubeStep = sqrt( step );
+		size_t cubeStep = sqrt( (float) step );
 		Pixel pixel;
 		
 		float red   = 0.0;
